Church_ProjectCopy/requirements.txt
asgiref==3.5.2
Django==4.1.4
django-crispy-forms==1.14.0
django-environ==0.9.0
django-filter==22.1
Pillow==9.3.0
psycopg2==2.9.5
sqlparse==0.4.3
tzdata==2022.7
